feat(replit): install @feathersjs/cli globally and delete upstream tsconfig.json
This commit adds the installation of @feathersjs/cli globally and deletes the upstream tsconfig.json file. Additionally, it updates the Node version used for debugging to v19.

This change adds TypeScript debugging with ts-node. To address a dap-node issue with CWD config, a helper script was added which changes the working dir with `--import` the ESM version of `-r`, new to Node v19.